consolidated top_five and main
